clean code, update delete, eidt video

+++ src/controllers/videoController.js
@@ -21,9 +21,15 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (video === null) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", {
     pageTitle: `Editing ${video.title} `,
@@ -32,19 +38,25 @@ export const getEdit = async (req, res) => {
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({
     _id: id,
   });
-
-  if (video === null) {
-    return res.status(404).render("404", { pageTitle: "Video not found." });
-  }
-  await Video.findByIdAndUpdate(id, {
+  const videoModified = await Video.findByIdAndUpdate(id, {
     title,
     description,
     hashtags: Video.formatHashtags(hashtags),
   });
+  if (video === null) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(videoModified.owner) !== String(video._id)) {
+    return res.status(403).redirect("/");
+  }
+
   return res.redirect(`/videos/${id}`);
 };
 
@@ -61,13 +73,16 @@ export const postUpload = async (req, res) => {
   const { title, description, hashtags } = req.body;
 
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl: fileUrl,
       owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -77,9 +92,21 @@ export const postUpload = async (req, res) => {
   }
 };
 export const deleteVideo = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (video === null) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
-
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
